requirements.txt
streamlit>=1.41.0
pandas>=2.0.0
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth>=2.0.0
plotly>=5.0.0
requests>=2.31.0